create a class for the mennu
